pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tcria"
version = "0.2.0"
description = "TCRIA — Legal evidence governance scanner (TCR gateway)."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Rodrigo Baptista da Silva"}]
dependencies = [
  "pypdf>=4.0.0",
  "reportlab>=4.0.0",
  "python-docx>=1.0.0",
  "streamlit>=1.30.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
tcria = "tcria.cli:main"

[tool.setuptools]
py-modules = [
  "run_governance_pipeline",
  "generate_blocked_artifacts_review",
  "generate_case_preparation_summary",
  "generate_case_timeline",
  "generate_unified_governance_report_pdf",
  "generate_investigation_report",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tcria*"]
